moved index.js to the src folder
