+++ package.json
@@ -2,7 +2,7 @@
   "name": "bit-bundler-splitter",
   "version": "6.0.1",
   "description": "bit-bundler plugin for splitting bundles up",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "eslint *.js && mocha --require babel-register"
   },
